LPS-128599 SF, inline the only usages
