+++ portal-test/src/com/liferay/portal/test/rule/LogAssertionAppender.java
@@ -23,8 +23,7 @@ import org.apache.log4j.spi.ThrowableInformation;
 
 /**
  * @author William Newbury
- * @deprecated As of Cavanaugh (7.4.x), replaced by {@link
- *             Log4jAssertionAppender}
+ * @deprecated As of Cavanaugh (7.4.x), with no direct replacement
  */
 @Deprecated
 public class LogAssertionAppender extends AppenderSkeleton {
