+++ portal-test/src/com/liferay/portal/test/rule/packageinfo
@@ -1,1 +1,1 @@
-version 3.3.0
+version 3.2.0
